refactor: restructure RunRow component into modular directory
- Convert single RunRow.tsx file to organized directory structure
- Split functionality into focused components:
  - RunRow.tsx: Main component logic
  - RunRowHeader.tsx: Header display logic
  - RunAnnotationRow.tsx: Individual annotation rendering
  - RunAnnotationsList.tsx: Annotations list management
  - SkeletonAnnotationRow.tsx: Loading state component
- Add supporting files for better organization:
  - types.ts: Component-specific type definitions
  - useRunAnnotations.ts: Custom hook for annotation logic
  - utils.ts: Utility functions
- Improve maintainability and code organization
